# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a ERP/POS API for G4H.",
        "title": "G4H ERP/POS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "G4H ERP/POS API",
            "url": "https://github.com/aslon1213/g4h_pos_erp",
            "email": "hamidovaslon1@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/admin/journals": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new journal entry for a branch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "journals"
                ],
                "summary": "Create a new journal entry",
                "parameters": [
                    {
                        "description": "New Journal Entry Input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.NewJournalEntryInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Output-models_JournalWithTransactionID"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorOutput"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/journals/branch/{branch_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Query journal entries by branch ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "journals"
                ],
                "summary": "Query journal entries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Branch ID",
                        "name": "branch_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Output-array_models_Journal"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorOutput"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/journals/report": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a report of journal entries",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "journals"
                ],
                "summary": "Get a report",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorOutput"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/journals/{journal_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a journal entry by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "journals"
                ],
                "summary": "Get a journal entry by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Journal ID",
                        "name": "journal_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Output-models_Journal"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorOutput"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/journals/{journal_id}/close": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Close a journal entry by updating its transactions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "journals"
                ],
                "summary": "Close a journal entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Journal ID",
                        "name": "journal_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Close Journal Entry Input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CloseJournalEntryInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Output-models_Journal"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorOutput"
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/journals/{journal_id}/reopen": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Reopen a journal entry by removing its closing transactions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "journals"
                ],
                "summary": "Reopen a closed journal entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Journal ID",
                        "name": "journal_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Output-models_Journal"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorOutput"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CloseJournalEntryInput": {
            "type": "object",
            "properties": {
                "cash_left": {
                    "type": "integer"
                },
                "terminal_income": {
                    "type": "integer"
                }
            }
        },
...
        "models.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.ErrorOutput": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {}
                },
                "error": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Error"
                    }
                }
            }
        },
...
        "models.JournalWithTransactionID": {
            "type": "object",
            "properties": {
                "branch": {
                    "$ref": "#/definitions/models.Branch"
                },
                "cash_left": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "operations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "shift_is_closed": {
                    "type": "boolean"
                },
                "terminal_income": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "models.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.NewJournalEntryInput": {
            "type": "object",
            "properties": {
                "branch_name_or_id": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                }
            }
        },
...
        "models.Output-array_models_Journal": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Journal"
                    }
                },
                "error": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Error"
                    }
                }
            }
        },
...
        "models.Output-models_Journal": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Journal"
                },
                "error": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Error"
                    }
                }
            }
        },
        "models.Output-models_JournalWithTransactionID": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.JournalWithTransactionID"
                },
                "error": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Error"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}